fix: refactor folders
