+++ lib/pages/home/names/details.dart
@@ -2,19 +2,6 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:ninety9names/repo/name.dart';
 
-List<Widget> displayName(Name n) {
-  Widget formatFields(String t, String st) {
-    return Expanded(child: ListTile(title: Text(t), subtitle: Text(st)));
-  }
-
-  return ([
-    formatFields('Number:', n.id!),
-    formatFields('Arabic:', n.arabic!),
-    formatFields('Transliteration:', n.transliteration!),
-    formatFields('Explanation:', n.explanation!),
-  ]);
-}
-
 class Details extends StatelessWidget {
   static const routeName = '/details';
 
@@ -32,4 +19,17 @@ class Details extends StatelessWidget {
       ),
     );
   }
+}
+
+List<Widget> displayName(Name n) {
+  Widget formatFields(String t, String st) {
+    return Expanded(child: ListTile(title: Text(t), subtitle: Text(st)));
+  }
+
+  return ([
+    formatFields('Number:', n.id!),
+    formatFields('Arabic:', n.arabic!),
+    formatFields('Transliteration:', n.transliteration!),
+    formatFields('Explanation:', n.explanation!),
+  ]);
 }
