+++ lib/pages/root/routes/routes.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:ninety9names/pages/home/home.dart';
-import 'package:ninety9names/pages/names/details.dart';
+import 'package:ninety9names/pages/home/names/details.dart';
 
 Map<String, Widget Function(BuildContext)> getRoutes() {
   return {
